Moved CleanUp back
